requirements.txt
llama-stack-client
streamlit
python-dotenv
requests
tavily-python
fire